refactor(core): move outcome grammar out of lib
