+++ devtools/build_topology_projection.py
@@ -119,7 +119,6 @@ LIB_ROOT_PRIMITIVES = frozenset(
         "__init__.py",
         "stats.py",
         "metrics.py",
-        "outcomes.py",
         "coverage.py",
         "repo_identity.py",
         "tail_overlay.py",
