+++ polylogue/core/outcomes.py
