+++ polylogue/showcase/invariants.py
@@ -12,7 +12,7 @@ from collections.abc import Callable
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from polylogue.lib.outcomes import OutcomeStatus
+from polylogue.core.outcomes import OutcomeStatus
 from polylogue.showcase.exercises import Exercise
 from polylogue.showcase.runner import ExerciseResult
