+++ tests/unit/core/test_audit.py
@@ -6,8 +6,8 @@ and AuditReport aggregation properties.
 
 from __future__ import annotations
 
-from polylogue.lib.outcomes import OutcomeCheck as CheckResult
-from polylogue.lib.outcomes import OutcomeStatus
+from polylogue.core.outcomes import OutcomeCheck as CheckResult
+from polylogue.core.outcomes import OutcomeStatus
 from polylogue.schemas.audit.checks import (
     check_annotation_coverage,
     check_privacy_guards,
