+++ tests/unit/proof/test_schema_provider_obligations.py
@@ -8,7 +8,7 @@ from hypothesis import strategies as st
 from hypothesis_jsonschema import from_schema
 
 from polylogue.core.json import JSONValue, json_document, require_json_value
-from polylogue.lib.outcomes import OutcomeStatus
+from polylogue.core.outcomes import OutcomeStatus
 from polylogue.proof.catalog import build_verification_catalog
 from polylogue.proof.models import ProofObligation
 from polylogue.proof.runners import (
